requirements.txt
alembic==1.6.5
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.66.0
greenlet==1.1.0
h11==0.12.0
idna==3.2
Mako==1.1.4
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.21
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0